course builder updated

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection") 
 
 //CREATE a new section
 exports.createSection = async(req,res) => {
@@ -79,7 +80,8 @@ exports.updateSection = async(req,res) => {
        .exec()
        res.status(200).json({
         success: true,
-        message: "Section Updated Successfully",
+        message: section,
+        data:course,
     });
     }catch(error) {
         console.error("Error updating section:", error);
@@ -93,18 +95,42 @@ exports.updateSection = async(req,res) => {
 //DELETE a section
 exports.deleteSection = async (req,res) => {
     try{
-         const {sectionId, courseId} = req.params;
+         const {sectionId, courseId} = req.body;
+         await Course.findByIdAndUpdate(courseId, {
+            $pull: {
+              courseContent: sectionId,
+            },
+          })
+        
+         const section = await Section.findById(sectionId)
+         console.log(sectionId, courseId)
+         if (!section) {
+           return res.status(404).json({
+             success: false,
+             message: "Section not found",
+           })
+         }
+
+         // Delete the associated subsections
+         await SubSection.deleteMany({ _id: { $in: section.subSection } })
+
+
          await Section.findByIdAndDelete(sectionId);
-         //Deleting sectionIs from course schema
-         const updatedCourseDetail = await Course.findByIdAndDelete (
-            courseId,
-            {sectionId},
-            {new:true}
-        );
-         res.status(200).json({
-            success: true,
-            message: "Section Deleted Successfully",
-        });
+
+          // find the updated course and return it
+        const course = await Course.findById(courseId)
+        .populate({
+        path: "courseContent",
+        populate: {
+            path: "subSection",
+        },
+        })
+        .exec()
+            res.status(200).json({
+                success: true,
+                message: "Section Deleted Successfully",
+                data:course,
+            });
     }catch(error) {
         console.error("Error deleting section:", error);
 		res.status(500).json({
